Constructor for the pgrid Class

Construct an object of class "pgrid" from a matrix or a higher-dimensional array.

Usage

pgrid(mass, boundary, gridtriple, generator, structure)

Arguments

mass

a matrix or higher-dimensional array specifing the masses in each pixel / at each pixel centre.

boundary, gridtriple, generator

arguments specifying the positions of the pixels. At most one of these can be specified.

structure

optional character string specifying the structure of the grid. Currently only "square" and "rectangular" make sense, and are derrived automatically from the dimensions of mass.

Details

For more detailed explanations of the arguments and other components of the derived object of class "pgrid", see pgrid-object.

Author

Dominic Schuhmacher dschuhm1@uni-goettingen.de

See also

Examples

m <- matrix(1:20, 4, 5)
a <- pgrid(m)
print(a)
#> Regularly spaced 4x5 pixel grid on [0,1] x [0,1.25].
#> x-gridtriple: 0.125 0.875 0.25 
#> y-gridtriple: 0.125 1.125 0.25 
#> pixel masses range from 1 to 20
#> total pixel mass: 210
#> total continuum mass: 13.125
print.default(a)
#> $structure
#> [1] "square"
#> 
#> $dimension
#> [1] 2
#> 
#> $n
#> [1] 4 5
#> 
#> $N
#> [1] 20
#> 
#> $boundary
#> [1] 0.00 1.00 0.00 1.25
#> 
#> $gridtriple
#>       [,1]  [,2] [,3]
#> [1,] 0.125 0.875 0.25
#> [2,] 0.125 1.125 0.25
#> 
#> $generator
#> $generator[[1]]
#> [1] 0.125 0.375 0.625 0.875
#> 
#> $generator[[2]]
#> [1] 0.125 0.375 0.625 0.875 1.125
#> 
#> 
#> $mass
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    1    5    9   13   17
#> [2,]    2    6   10   14   18
#> [3,]    3    7   11   15   19
#> [4,]    4    8   12   16   20
#> 
#> $totmass
#> [1] 210
#> 
#> $totcontmass
#> [1] 13.125
#> 
#> attr(,"class")
#> [1] "pgrid"

if (FALSE) {  
  plot(a, rot=TRUE)}