Return Text Strings for lbfgs Return Codes

Given a vector of return codes, give back the corresponding vector of return strings from the lbfgs library. Nonexistant codes are ignored.

Usage

ret_message(n = NULL)

Arguments

n

The vector of return codes or NULL meaning the whole list shall be returned.

Value

A named character vector of the corresponding return strings.

Note

Code 0 is ignored, since for technical reasons it is never returned by the function semidiscrete1.

See also

Author

Dominic Schuhmacher schuhmacher@math.uni-goettingen.de

Examples

ret_message()
#>                                    1                                    2 
#>                         "LBFGS_STOP"            "LBFGS_ALREADY_MINIMIZED" 
#>                                -1024                                -1023 
#>              "LBFGSERR_UNKNOWNERROR"                "LBFGSERR_LOGICERROR" 
#>                                -1022                                -1021 
#>               "LBFGSERR_OUTOFMEMORY"                  "LBFGSERR_CANCELED" 
#>                                -1020                                -1019 
#>                 "LBFGSERR_INVALID_N"             "LBFGSERR_INVALID_N_SSE" 
#>                                -1018                                -1017 
#>             "LBFGSERR_INVALID_X_SSE"           "LBFGSERR_INVALID_EPSILON" 
#>                                -1016                                -1015 
#>        "LBFGSERR_INVALID_TESTPERIOD"             "LBFGSERR_INVALID_DELTA" 
#>                                -1014                                -1013 
#>        "LBFGSERR_INVALID_LINESEARCH"           "LBFGSERR_INVALID_MINSTEP" 
#>                                -1012                                -1011 
#>           "LBFGSERR_INVALID_MAXSTEP"              "LBFGSERR_INVALID_FTOL" 
#>                                -1010                                -1009 
#>             "LBFGSERR_INVALID_WOLFE"              "LBFGSERR_INVALID_GTOL" 
#>                                -1008                                -1007 
#>              "LBFGSERR_INVALID_XTOL"     "LBFGSERR_INVALID_MAXLINESEARCH" 
#>                                -1006                                -1005 
#>       "LBFGSERR_INVALID_ORTHANTWISE" "LBFGSERR_INVALID_ORTHANTWISE_START" 
#>                                -1004                                -1003 
#>   "LBFGSERR_INVALID_ORTHANTWISE_END"             "LBFGSERR_OUTOFINTERVAL" 
#>                                -1002                                -1001 
#>         "LBFGSERR_INCORRECT_TMINMAX"            "LBFGSERR_ROUNDING_ERROR" 
#>                                -1000                                 -999 
#>               "LBFGSERR_MINIMUMSTEP"               "LBFGSERR_MAXIMUMSTEP" 
#>                                 -998                                 -997 
#>         "LBFGSERR_MAXIMUMLINESEARCH"          "LBFGSERR_MAXIMUMITERATION" 
#>                                 -996                                 -995 
#>             "LBFGSERR_WIDTHTOOSMALL"         "LBFGSERR_INVALIDPARAMETERS" 
#>                                 -994 
#>          "LBFGSERR_INCREASEGRADIENT" 
ret_message(c(2,-1023,-1019))
#>                         2                     -1023                     -1019 
#> "LBFGS_ALREADY_MINIMIZED"     "LBFGSERR_LOGICERROR"  "LBFGSERR_INVALID_N_SSE"