Optimal Transport in Various Forms

Solve optimal transport problems. Compute Wasserstein distances (a.k.a. Kantorovitch, Fortet--Mourier, Mallows, Earth Mover's, or minimal \(L_p\) distances), return the corresponding transport plans, and display them graphically. Objects that can be compared include grey-scale images, (weighted) point patterns, and mass vectors.

Details

Package:	transport

Type:	Package

Version:	0.12-1

Date:	2019-08-07

License:	GPL (>=2)

LazyData:	yes

The main end-user function is transport. It computes optimal transport plans between images (class pgrid), point patterns (class pp), weighted point patterns (class wpp) and mass vectors, based on various algorithms. These transport plans can be ploted. The function wasserstein allows for the numerical computation of \(p\)-th order Wasserstein distances.

Most functions in this package are designed for data in two and higher dimensions. A quick tool for computing the \(p\)-th order Wasserstein distance between univariate samples is wasserstein1d.

Author

Dominic Schuhmacher schuhmacher@math.uni-goettingen.de

Björn Bähre bjobae@gmail.com (code for aha-method)

Nicolas Bonneel nicolas.bonneel@liris.cnrs.fr
(adaptation of LEMON code for fast networkflow method)

Carsten Gottschlich gottschlich@math.uni-goettingen.de
(original java code for shortlist and revsimplex methods)

Valentin Hartmann valentin.hartmann@epfl.ch (code for aha method for p=1)

Florian Heinemann florian.heinemann@uni-goettingen.de
(integration of networkflow method)

Bernhard Schmitzer schmitzer@uni-muenster.de (shielding method)

Jörn Schrieber joern.schrieber-1@mathematik.uni-goettingen.de (subsampling method)

Maintainer: Dominic Schuhmacher dominic.schuhmacher@mathematik.uni-goettingen.de

References

See help page for the function transport.

Examples

##  See examples for function transport