Set the Control Parameters Used by transport.

Set the control parameters for the algorithm used by the function transport.

Usage

trcontrol(method = c("networkflow", "revsimplex", "shortsimplex", "primaldual", "aha",
          "shielding", "auction", "auctionbf"), para = list(),
          start = c("auto",          "modrowmin", "nwcorner", "russell"), 
          nscales = 1, scmult = 2,returncoarse = FALSE, a = NULL,
          b = NULL, M = NULL, N = NULL)

Arguments

method

The algorithm to be used to compute the optimal transference plan. See details for the function transport.pgrid.

para

A list of parameters that are specific to the chosen method. See the table on the help page of the function transport.

start

If method == "revsimplex", the method for computing a starting solution.

nscales, scmult, returncoarse

The parameters for the multiscale versions of certain algorithms. See the help on transport.

a,b,M,N

The two objects a and b for which the transportation problem is to be solved or the sizes M and N of these objects. Based on the information available here, trcontrol tries hard to find reasonable values for the control parameters of the algorithm not specified directly.

Details

For further details about the parameters of the individual algorithms see the help page for transport.

Value

A list with components method, para, start, nscales, scmult, returncoarse as entered or adapted/computed based on the arguments method, a, b, M, N.

Author

Dominic Schuhmacher dschuhm1@uni-goettingen.de

Note

This function is typically only called by the user to check what the parameter settings used by the function transport are for a given problem.

See also