Constructor for the wpp Class

Construct an object of class "wpp" from a matrix of points and a vector of masses.

Usage

wpp(coordinates, mass)

Arguments

coordinates

a matrix specifying the coordinates of the points. Each row corresponds to a point.

mass

a vector of non-negative values specifying the masses at these points.

Details

For more detailed explanations of the arguments and other components of the returned object of class "wpp", see wpp-object.

It is legitimate to assign mass 0 to individual points in the arguments. However, when constructing the wpp-object such points are deleted. The coordinates of the deleted points can still be accessed via the attribute zeropoints.

Author

Dominic Schuhmacher dschuhm1@uni-goettingen.de
Timo Wilm timo.wilm@stud.uni-goettingen.de

See also

Examples

m <- matrix(c(1,1,2,2,3,1,4,2),4,2)
a <- pp(m)
print(a)
#> Pattern of 4 points in 2 dimensions.
#> Minimal coordinates: 1 1 
#> Maximal coordinates: 2 4 
print.default(a)
#> $dimension
#> [1] 2
#> 
#> $N
#> [1] 4
#> 
#> $coordinates
#>      [,1] [,2]
#> [1,]    1    3
#> [2,]    1    1
#> [3,]    2    4
#> [4,]    2    2
#> 
#> attr(,"class")
#> [1] "pp"

if (FALSE) {  
  plot(a)}